NIT Move SemanticDB-related integration tests to a dedicated trait
